backend/requirements.txt
python-docx
easyocr
django 
djangorestframework
pypdf
langchain_openai